+++ src/pages/policyPage.js
@@ -70,8 +70,8 @@ export default function PolicyPage({ getTenants, tenantValues, thisTenant, graph
         .get((typeof env !== 'undefined' ? env.ANUBIS_API_URL : '') + 'v1/policies', {
           headers: {
             'fiware-service': GeTenantData('name'),
-            'fiware-servicepath': service.path
-            //'authorization': `Bearer ${token}`
+            'fiware-servicepath': service.path,
+            authorization: `Bearer ${token}`
           }
         })
         .then((response) => {
@@ -106,8 +106,8 @@ export default function PolicyPage({ getTenants, tenantValues, thisTenant, graph
         .get((typeof env !== 'undefined' ? env.ANUBIS_API_URL : '') + 'v1/policies' + queryParameters, {
           headers: {
             'fiware-service': GeTenantData('name'),
-            'fiware-servicepath': policyFilter !== null ? policyFilter.fiware_service_path : service.path
-            //'authorization': `Bearer ${token}`
+            'fiware-servicepath': policyFilter !== null ? policyFilter.fiware_service_path : service.path,
+            authorization: `Bearer ${token}`
           }
         })
         .then((response) => {
@@ -141,7 +141,7 @@ export default function PolicyPage({ getTenants, tenantValues, thisTenant, graph
       axios
         .get((typeof env !== 'undefined' ? env.ANUBIS_API_URL : '') + 'v1/policies/access-modes', {
           headers: {
-            //'authorization': `Bearer ${token}`
+            authorization: `Bearer ${token}`
           }
         })
         .then((response) => setAccess_modes(response.data))
@@ -159,7 +159,7 @@ export default function PolicyPage({ getTenants, tenantValues, thisTenant, graph
       axios
         .get((typeof env !== 'undefined' ? env.ANUBIS_API_URL : '') + 'v1/policies/agent-types', {
           headers: {
-            //'authorization': `Bearer ${token}`
+            authorization: `Bearer ${token}`
           }
         })
         .then((response) => setagentsTypes(response.data))