+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - Add support for retrieving resource id based
   on resource type endpoint in policy forms
 - Update tests to use Node 16 (#93)
+- Pass authorization token to policies calls
 
 ### Bug fixes